VithallDjango/requirements.txt
Django==4.0.5
docutils
Pillow
django-import-export
django-pure-pagination